fix: separate API from implementation
