+++ auth/credentials/oauth_client.go
@@ -12,9 +12,7 @@ import (
 	"time"
 )
 
-// Token represents the OAuth2 token structure
-// Note token is internal to the SDK and should never be exposed outside sdk.
-// For more information see TokenSource implementation
+// Token represents the internal OAuth2 token structure
 type token struct {
 	AccessToken string    `json:"access_token"`
 	Expiry      time.Time `json:"expiry,omitempty"`
@@ -66,7 +64,7 @@ func (c *OAuthClient) refreshToken() (*token, error) {
 	return newToken, nil
 }
 
-// fetchToken makes a manual POST request to the OAuth server to get the access token.
+// fetchToken makes a manual POST request to Server (tokenUrl) to fetch the access token.
 func (c *OAuthClient) fetchToken() (*token, error) {
 	data := url.Values{}
 	data.Set("grant_type", "client_credentials")
@@ -113,6 +111,7 @@ func (t *token) SetAuthHeader(r *http.Request) {
 	r.Header.Set("Authorization", "Bearer "+t.AccessToken)
 }
 
+// Additional time for Access Tokens to not expire.
 const expiryDelta = 10 * time.Second
 
 // expired checks if the token is close to expiring.
@@ -123,7 +122,7 @@ func (t *token) expired() bool {
 	return t.Expiry.Round(0).Add(-expiryDelta).Before(time.Now())
 }
 
-// Valid checks if the token is still valid.
+// Valid checks if the token is still valid (present and not expired)
 func (t *token) Valid() bool {
 	return t != nil && t.AccessToken != "" && !t.expired()
 }
@@ -132,7 +131,7 @@ func (t *token) Valid() bool {
 func parseToken(accessToken string) (*token, error) {
 	parts := strings.Split(accessToken, ".")
 	if len(parts) != 3 {
-		return nil, errors.New("invalid token format")
+		return nil, errors.New("invalid access token format")
 	}
 
 	payload, err := base64.RawURLEncoding.DecodeString(parts[1])
@@ -157,59 +156,4 @@ func parseToken(accessToken string) (*token, error) {
 		Expiry:      expiry,
 		ExpiresIn:   int(time.Until(expiry).Seconds()),
 	}, nil
-}
-
-// NewHTTPClientWithServiceAccountAuth creates an HTTP client with the custom transport.
-func NewHTTPClientWithServiceAccountAuth(client *OAuthClient) *http.Client {
-	return &http.Client{
-		Transport: &OAuthCustomHTTPTransport{
-			underlyingTransport: http.DefaultTransport,
-			client:              client,
-		},
-	}
-}
-
-// NewServiceAccountOAuthClient initializes an OAuthClient with client credentials with default InMemoryTokenSource.
-func NewServiceAccountOAuthClient(clientID, clientSecret string) *OAuthClient {
-	return NewServiceAccountOAuthClientWithTokenSource(ServiceAccountOAuthClientWithTokenSource{
-		ClientID:     clientID,
-		ClientSecret: clientSecret,
-		TokenSource:  &InMemoryTokenSource{},
-	})
-}
-
-type ServiceAccountOAuthClientWithTokenSource struct {
-	ClientID     string
-	ClientSecret string
-	// Custom token source. InMemoryTokenSource being default
-	TokenSource TokenSource
-
-	// Custom context
-	Context *context.Context
-	// Custom base url for change
-	BaseURL *string
-}
-
-// NewServiceAccountOAuthClient initializes an OAuthClient with client credentials and custom TokenSource
-// Use this method to initialze custom token storage (filesystem)
-func NewServiceAccountOAuthClientWithTokenSource(opts ServiceAccountOAuthClientWithTokenSource) *OAuthClient {
-	//nolint:gosec //url only
-	tokenURL := "https://cloud.mongodb.com/api/oauth/token"
-	if opts.BaseURL != nil {
-		tokenURL = *opts.BaseURL + "/api/oauth/token"
-	}
-	var ctx context.Context
-	if opts.Context == nil {
-		ctx = context.Background()
-	} else {
-		ctx = *opts.Context
-	}
-
-	return &OAuthClient{
-		clientID:     opts.ClientID,
-		clientSecret: opts.ClientSecret,
-		tokenURL:     tokenURL,
-		tokenSource:  opts.TokenSource,
-		ctx:          ctx,
-	}
 }
